ClassExample_WorksheetExample
Going over student question on the worksheet problems. Start
Calculate the discounted price for an item. Create an expression that
will calculate the discounted price with and without sales tax. (It is
acceptable for the result to have more than two digits after the
decimal. $345.896 for example.) Givens: Original price Discount
percentage (20% would be 20, for example) Description of item Sales tax
percentage Result Variables: Price of the item with tax Price of the
item without tax Result to Print: “Your X was originally $X, but after
a X% discount, it is now $X without tax, and $X with tax.”